requirements.txt
python-dotenv
flask
ruamel.yaml
pyVim
pyVmomi
